fix změny a odstranění úkolů proti škůdcům

+++ odstranitUkol.php
@@ -13,10 +13,20 @@ if(empty($currentUser) || ($currentUser['role']!='admin')){
 }
 
 if (!empty($_GET['id'])){
-    $query = $db->prepare('DELETE FROM posts WHERE post_id=:id');
-    $query->execute([
-        ":id"=>$_GET['id']
-    ]);
+
+    $postQuery= $db->prepare('SELECT * FROM posts JOIN users USING (user_id) WHERE post_id=:id LIMIT 1;');
+    $postQuery->execute([':id'=>$_GET['id']]);
+    if ($post=$postQuery->fetch(PDO::FETCH_ASSOC)){
+        if($post['family_id']!=$_SESSION['family_id']) {
+            die('Tento úkol nebyl nalezen.');
+        }
+        $query = $db->prepare('DELETE FROM posts WHERE post_id=:id');
+        $query->execute([
+            ":id"=>$_GET['id']
+        ]);
+    }else{
+        exit('Tento úkol neexistuje.');//tady by mohl být i lepší výpis chyby :)
+    }
 }
 
 header('Location: index.php');
